refactor: extract skill execution ownership
